requirements.txt
click
steam
discord.py